requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
codemetrics==0.11.7
colorama==0.4.6
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
joblib==1.2.0
lizard==1.17.10
mando==0.6.4
mypy-extensions==0.4.3
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
radon==5.1.0
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tqdm==4.64.1
urllib3==1.26.12
